QR Code Generator/requirements.txt
qrcode[pil]
pillow



